+++ CHANGELOG.md
@@ -20,3 +20,10 @@
 ## 1.5.2+4
 * Pub dev warnings fix
 * Post, Put and Delete Http Requests Minor bug fix
+
+## 1.5.3+5
+* Fixed return types on io devices
+* Added a structured ActiveResponse data, endpoint, errors
+* Now errors will be returned in ActiveResponse data
+* Http put error management, post error management and delete requests error management updated.
+* Simple uuid generator.